requirements.txt
fastapi # веб фреймворк fastapi
uvicorn # ASGI веб сервер
pydantic # для серилизация типов
python-dotenv # для переменных виртуального окружения
fastapi-sqlalchemy # для работы с бд
alembic # для миграций
python-jose # для работы с JWT
passlib[bcrypt] # для хешировани и валидации паролей
python-multipart # парсер, используется для форм
requests # for fastapi pytest
psycopg2-binary # адаптер для работы с PostgreSQL
httpx